requirements.txt
absl-py==2.0.0
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.0
aniso8601==9.0.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
asciitree==0.3.3
asteroid-filterbanks==0.4.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
# av==10.0.0
beautifulsoup4==4.12.2
blinker==1.7.0
boto3==1.33.8
botocore==1.33.8
braceexpand==0.1.7
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
Cython
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.0
comm==0.2.0
contourpy==1.2.0
ctranslate2==3.23.0
cycler==0.12.1
datasets==2.15.0
decorator==5.1.1
deepmultilingualpunctuation==1.0.1
dill==0.3.7
Distance==0.1.3
docopt==0.6.2
editdistance==0.6.2
einops==0.7.0
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.7.4
fastapi==0.104.1
fasteners==0.19
faster-whisper==0.10.0
fasttext==0.9.2
filelock==3.13.1
Flask==3.0.0
Flask-RESTful==0.3.10
flatbuffers==23.5.26
fonttools==4.46.0
frozenlist==1.4.0
fsspec==2023.10.0
ftfy==6.1.3
g2p-en==2.1.0
gdown==4.7.1
google-auth==2.25.1
google-auth-oauthlib==1.1.0
greenlet==3.0.1
grpcio==1.59.3
h11==0.14.0
h5py==3.10.0
httptools==0.6.1
huggingface-hub==0.19.4
humanfriendly==10.0
hydra-core==1.3.2
HyperPyYAML==1.2.2
idna==3.6
ijson==3.2.3
inflect==7.0.0
install==1.3.5
ipython==8.18.1
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.2
jiwer==3.0.3
jmespath==1.0.1
joblib==1.3.2
julius==0.2.7
jupyterlab-widgets==3.0.9
kaldi-python-io==1.2.2
kaldiio==2.18.0
kiwisolver==1.4.5
lazy_loader==0.3
Levenshtein==0.23.0
librosa==0.10.1
lightning==2.1.2
lightning-utilities==0.10.0
llvmlite==0.41.1
loguru==0.7.2
lxml==4.9.3
Mako==1.3.0
Markdown==3.5.1
markdown-it-py==3.0.0
markdown2==2.4.11
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
megatron-core==0.3.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.15
nemo-toolkit @ git+https://github.com/NVIDIA/NeMo.git@52d50e9e09a3e636d60535fd9882f3b3f32f92ad
networkx==3.2.1
nltk==3.8.1
numba==0.58.1
numcodecs==0.12.1
numpy==1.23.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.16.3
OpenCC==1.1.6
optuna==3.4.0
packaging==23.2
pandas==2.1.3
pangu==4.0.6.1
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
plac==1.4.1
platformdirs==4.1.0
pooch==1.8.0
portalocker==2.8.2
primePy==1.3
prompt-toolkit==3.0.41
protobuf==4.23.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyannote.audio==3.1.0
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==14.0.1
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybind11==2.11.1
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytorch-lightning==2.1.2
pytorch-metric-learning==2.3.0
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.5.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rouge-score==0.1.2
rsa==4.9
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
setuptools
s3transfer==0.8.2
sacrebleu==2.3.3
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
semver==3.0.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
sox==1.4.1
soxr==0.3.7
speechbrain==0.5.16
SQLAlchemy==2.0.23
stack-data==0.6.3
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorstore==0.1.45
termcolor==2.4.0
text-unidecode==1.3
texterrors==0.4.4
threadpoolctl==3.2.0
tokenizers==0.15.0
torch==2.1.1
torch-audiomentations==0.11.0
torch-pitch-shift==1.2.4
torchaudio==2.1.1
torchmetrics==1.2.1
torchvision==0.16.1
tqdm==4.66.1
traitlets==5.14.0
transformers==4.35.2
triton==2.1.0
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.12
webdataset==0.2.85
websockets==12.0
Werkzeug==3.0.1
wget==3.2
whisper==1.1.10
whisperx @ git+https://github.com/m-bain/whisperx.git
widgetsnbextension==4.0.9
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.3
youtokentome
zarr==2.16.1